Split oidcprovider into client and providers
