+++ client/providers/mocks.go
@@ -1,4 +1,4 @@
-package oidcprovider
+package providers
 
 import (
 	"context"
@@ -9,6 +9,7 @@ import (
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jwt"
 	"github.com/lestrrat-go/jwx/v2/jwt/openid"
+	"github.com/openpubkey/openpubkey/client"
 	"github.com/openpubkey/openpubkey/util"
 )
 
@@ -56,7 +57,7 @@ func (m *MockOpenIdProvider) RequestTokens(ctx context.Context, cicHash string) 
 	return signedToken, nil
 }
 
-func (m *MockOpenIdProvider) PublicKey(ctx context.Context, idt []byte) (PublicKey, error) {
+func (m *MockOpenIdProvider) PublicKey(ctx context.Context, idt []byte) (client.PublicKey, error) {
 	return m.signer.Public().(*rsa.PublicKey), nil
 }
